# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Fullcycle Pós Go Expert Go Expert",
        "title": "Desafio 2.0 - Multithreading",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Giovani Angelo",
            "email": "giovani.angelo@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/cep/{cep}": {
            "get": {
                "description": "Get Address by Post Code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "address"
                ],
                "summary": "Get Address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cep",
                        "name": "cep",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.AnyAddress"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.AnyAddress": {
            "type": "object",
            "properties": {
                "api": {
                    "type": "string"
                },
                "bairro": {
                    "type": "string"
                },
                "cep": {
                    "type": "string"
                },
                "localidade": {
                    "type": "string"
                },
                "logradouro": {
                    "type": "string"
                },
                "uf": {
                    "type": "string"
                }
            }
        }
    }
}